Cambios

+++ src/main/java/acme/forms/DeveloperDashboard.java
@@ -15,12 +15,12 @@ public class DeveloperDashboard extends AbstractForm {
 
 	private int					totalTrainingSessionsWithLink;
 
-	private double				averageTrainingModuleTime;
+	private Double				averageTrainingModuleTime;
 
-	private double				deviationTrainingModuleTime;
+	private Double				deviationTrainingModuleTime;
 
-	private int					minimumTrainingModuleTime;
+	private Double				minimumTrainingModuleTime;
 
-	private int					maximumTrainingModuleTime;
+	private Double				maximumTrainingModuleTime;
 
 }
